backend/requirements.txt
Django==4.2.26
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.3.1
requests==2.31.0
gunicorn==22.0.0
